# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Nfc locks api",
        "description": "Nfc locks",
        "contact": {
            "email": "tcom.developer@gmail.com"
        },
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "http://messenger.test/api",
            "description": "Dev server"
        }
    ],
    "paths": {
        "/auth/register": {
            "post": {
                "tags": [
                    "Auth"
                ],
                "summary": "Register new user",
                "operationId": "ed42f392010c301892e640067141676e",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "email",
                                    "password"
                                ],
                                "properties": {
                                    "name": {
                                        "description": "user name",
                                        "type": "string",
                                        "default": "newuser"
                                    },
                                    "email": {
                                        "description": "email",
                                        "type": "string",
                                        "default": "newuser@mail.com"
                                    },
                                    "password": {
                                        "description": "password",
                                        "type": "string",
                                        "default": "Secret123#"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "User registered"
                    },
                    "500": {
                        "description": "Server Error"
                    }
                }
            }
        },
        "/auth/login": {
            "post": {
                "tags": [
                    "Auth"
                ],
                "summary": "Login user",
                "operationId": "8cb7742ae98d0990acb2907c4e00bf24",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "email",
                                    "password"
                                ],
                                "properties": {
                                    "email": {
                                        "description": "email",
                                        "type": "string",
                                        "default": "user@mail.com"
                                    },
                                    "password": {
                                        "description": "password",
                                        "type": "string",
                                        "default": "Secret123#"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Korisnik prijavljen"
                    },
                    "500": {
                        "description": "Server Error"
                    }
                }
            }
        },
        "/auth/whoami": {
            "post": {
                "tags": [
                    "Auth"
                ],
                "summary": "Podaci o korisniku na osnovu tokena",
                "operationId": "7cd7c8c9f7981009260efca1565a769b",
                "responses": {
                    "200": {
                        "description": "Podaci o korisniku vraćeni na front"
                    },
                    "500": {
                        "description": "Server Error"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/auth/logout": {
            "post": {
                "tags": [
                    "Auth"
                ],
                "summary": "Odjava sa sistema",
                "operationId": "0d29c740dc06d77602af8095bf4654cc",
                "responses": {
                    "200": {
                        "description": "Korisnik odjavljen iz sistema"
                    },
                    "500": {
                        "description": "Server Error"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/connection/initiate": {
            "post": {
                "tags": [
                    "Connection"
                ],
                "summary": "Initiate new connection",
                "operationId": "9fb490977767a4d0f9f4ae7da304f404",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "recipient_id"
                                ],
                                "properties": {
                                    "recipient_id": {
                                        "type": "integer",
                                        "default": 3
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Connection Initiated"
                    },
                    "500": {
                        "description": "Server Error"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/connection/accept": {
            "post": {
                "tags": [
                    "Connection"
                ],
                "summary": "Accept Connection",
                "operationId": "735cb7e734238cf2c0df1c96fd8818fc",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "connection_id"
                                ],
                                "properties": {
                                    "connection_id": {
                                        "type": "integer",
                                        "default": 1
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Connection accepted"
                    },
                    "500": {
                        "description": "Server Error"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/connection/reject": {
            "post": {
                "tags": [
                    "Connection"
                ],
                "summary": "Reject Connection",
                "operationId": "c7293bdddb4b0866d7532558b788cb58",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "connection_id"
                                ],
                                "properties": {
                                    "connection_id": {
                                        "type": "integer",
                                        "default": 1
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Connection rejected"
                    },
                    "500": {
                        "description": "Server Error"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/user/search": {
            "post": {
                "tags": [
                    "User"
                ],
                "summary": "Search user by name or email",
                "operationId": "7495095e5b6c23625c419daaa0200f1a",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "search"
                                ],
                                "properties": {
                                    "search": {
                                        "description": "search string",
                                        "type": "string",
                                        "default": "user"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "User registered"
                    },
                    "500": {
                        "description": "Server Error"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/user/show/{id}": {
            "get": {
                "tags": [
                    "User"
                ],
                "summary": "Find user by ID",
                "operationId": "2359a38d992595bc4001f41c0b92d89d",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "User ID",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "User found"
                    },
                    "500": {
                        "description": "Server Error"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        }
    },
    "components": {
        "schemas": {
            "Controller": {
                "format": "https"
            }
        },
        "securitySchemes": {
            "bearerAuth": {
                "type": "http",
                "description": "Authorize with bearer token",
                "name": "Authorization",
                "in": "header",
                "scheme": "Bearer"
            }
        }
    },
    "tags": [
        {
            "name": "Auth",
            "description": "Auth"
        },
        {
            "name": "Connection",
            "description": "Connection"
        },
        {
            "name": "User",
            "description": "User"
        }
    ]
}